+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/FakeClock.java
@@ -182,7 +182,8 @@ public class FakeClock implements Clock {
       // This isn't a looper message created by this class, so no need to handle the blocking.
       return;
     }
-    busyLoopers.add(checkNotNull(Looper.myLooper()));
+    busyLoopers.add(currentLooper);
+    ThreadTestUtil.unblockThreadsWaitingForProgressOnCurrentLooper();
     waitingForMessage = false;
     maybeTriggerMessage();
   }