+++ libraries/test_utils_robolectric/src/main/java/androidx/media3/test/utils/robolectric/RobolectricUtil.java
@@ -22,6 +22,7 @@ import androidx.media3.common.util.Clock;
 import androidx.media3.common.util.ConditionVariable;
 import androidx.media3.common.util.SystemClock;
 import androidx.media3.common.util.UnstableApi;
+import androidx.media3.test.utils.ThreadTestUtil;
 import com.google.common.base.Supplier;
 import java.util.concurrent.TimeoutException;
 import org.robolectric.shadows.ShadowLooper;
@@ -116,6 +117,7 @@ public final class RobolectricUtil {
     if (Looper.myLooper() != looper) {
       throw new IllegalStateException();
     }
+    ThreadTestUtil.unblockThreadsWaitingForProgressOnCurrentLooper();
     ShadowLooper shadowLooper = shadowOf(looper);
     long timeoutTimeMs = clock.currentTimeMillis() + timeoutMs;
     while (!condition.get()) {