+++ libraries/test_utils_robolectric/src/test/java/androidx/media3/test/utils/robolectric/RobolectricUtilTest.java
@@ -23,8 +23,10 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import android.os.Looper;
 import androidx.media3.common.util.Clock;
 import androidx.media3.common.util.ConditionVariable;
+import androidx.media3.test.utils.ThreadTestUtil;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.common.base.Supplier;
 import java.util.concurrent.TimeoutException;
@@ -89,5 +91,36 @@ public class RobolectricUtilTest {
     RobolectricUtil.runMainLooperUntil(mockCondition, /* timeoutMs= */ 5674, mock(Clock.class));
 
     verify(mockCondition, times(5)).get();
+  }
+
+  @Test
+  public void
+      runMainLooperUntil_whenConditionIsBlockedOnOtherThreadWaitingForProgress_unblocksItself()
+          throws Exception {
+    Clock mockClock = mock(Clock.class);
+    ConditionVariable testCondition = new ConditionVariable();
+    ConditionVariable testThreadReady = new ConditionVariable();
+    Thread thread =
+        new Thread("RobolectricUtilsTest") {
+          @Override
+          public void run() {
+            ConditionVariable blockedCondition = new ConditionVariable();
+            ThreadTestUtil.registerThreadIsBlockedUntilProgressOnLooper(
+                blockedCondition, Looper.getMainLooper());
+            testThreadReady.open();
+            try {
+              blockedCondition.block();
+            } catch (InterruptedException e) {
+              // Ignore.
+            }
+            testCondition.open();
+          }
+        };
+    thread.start();
+    testThreadReady.block();
+
+    // Verify the thread gets unblocked.
+    RobolectricUtil.runMainLooperUntil(testCondition::isOpen, /* timeoutMs= */ 42, mockClock);
+    thread.join();
   }
 }